Convert Game, Board, and GameOver from class to function.
They all use componentWillMount to update the current screen in the model, so that the follower view will change when the leader view renders a new page. The timing isn't exactly the same as useEffect, but it doesn't matter.

Question remains since it is using componentWillReceiveProps to compute state based on props.