Clean up state management.
The Scores and Question view were using state for the leader view and props for the follower view which was confusing. Furthermore the polling for updates was always a code smell. This change uses the BroadcastChannel API to notify the follower view of updates and uses a custom hook such that the game state automatically updates whenever any service mutation method is called regardless of leader or follower.